Move accommodation referral history files
We move our accommodation referral history files for consistency
