+++ cypress_shared/pages/apply/accommodation-need/behaviour-in-previous-accommodation/previousStays.ts
