+++ cypress_shared/pages/apply/accommodation-need/behaviour-in-previous-accommodation/previousStaysDetails.ts
