+++ cypress_shared/pages/apply/index.ts
@@ -1,5 +1,5 @@
-import PreviousStaysDetailsPage from './accommodation-need/accommodation-referral-history/referralHistoryDetails'
-import PreviousStaysPage from './accommodation-need/accommodation-referral-history/referralsPreviouslySubmitted'
+import PreviousStaysPage from './accommodation-need/behaviour-in-previous-accommodation/previousStays'
+import PreviousStaysDetailsPage from './accommodation-need/behaviour-in-previous-accommodation/previousStaysDetails'
 import ConsentDetailsPage from './accommodation-need/consent/consentDetails'
 import ConsentGivenPage from './accommodation-need/consent/consentGiven'
 import BackupContactPage from './accommodation-need/contact-details/backupContact'
